% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_descritiva.R
\name{analise_descritiva}
\alias{analise_descritiva}
\title{Resultados para relat칩rio de an치lise descritiva}
\usage{
analise_descritiva(dados)
}
\description{
Resultados para relat칩rio de an치lise descritiva
}
