% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_univariada.R
\name{analise_univariada}
\alias{analise_univariada}
\title{An치lise univariada automatizada para relat칩rio.}
\usage{
analise_univariada(variavel, cor = "#028090")
}
\description{
An치lise univariada automatizada para relat칩rio.
}
