% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_bivar_info.R
\name{dataset_bivar_info}
\alias{dataset_bivar_info}
\title{Qual teste utlizar em cada par de variáveis de um dataset}
\usage{
dataset_bivar_info(dados)
}
\description{
Qual teste utlizar em cada par de variáveis de um dataset
}
