% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_normalidade.R
\name{dataset_normalidade}
\alias{dataset_normalidade}
\title{Teste de normalidade para todos as variaveis do dataset:}
\usage{
dataset_normalidade(dados, ...)
}
\description{
Teste de normalidade para todos as variaveis do dataset:
}
