% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_tabela_dataset_estrutura.R
\name{eda_tabela_dataset_estrutura}
\alias{eda_tabela_dataset_estrutura}
\title{Estrutura dos dados
Data de atualização 23/09/2022}
\usage{
eda_tabela_dataset_estrutura(dados)
}
\description{
Estrutura dos dados
Data de atualização 23/09/2022
}
