% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_tabela_dataset_geral.R
\name{eda_tabela_dataset_geral}
\alias{eda_tabela_dataset_geral}
\title{Visão geral de banco de dados
Retorna as seguintes informações de um banco de dados:**
- Nome das variáveis
- Classe das variáveis
- Normalidade (caso numeric)
- Niveis e frequencia (casso character ou factor)
- Numero de observações não ausentes em cada variável}
\usage{
eda_tabela_dataset_geral(dados)
}
\arguments{
\item{dados}{banco de dados}
}
\description{
Visão geral de banco de dados
Retorna as seguintes informações de um banco de dados:**
- Nome das variáveis
- Classe das variáveis
- Normalidade (caso numeric)
- Niveis e frequencia (casso character ou factor)
- Numero de observações não ausentes em cada variável
}
