% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_tabela_dataset_resumo_num.R
\name{eda_tabela_dataset_resumo_num}
\alias{eda_tabela_dataset_resumo_num}
\title{Resumo das variáveis numéricas}
\usage{
eda_tabela_dataset_resumo_num(dados)
}
\description{
Resumo das variáveis numéricas
}
