% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_tabela_dataset_teste_qui_quadrado.R
\name{eda_tabela_dataset_teste_qui_quadrado}
\alias{eda_tabela_dataset_teste_qui_quadrado}
\title{Testes qui-quadrado para todas as variáveis catégoricas}
\usage{
eda_tabela_dataset_teste_qui_quadrado(dados, resposta = NULL)
}
\description{
Testes qui-quadrado para todas as variáveis catégoricas
}
