% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_visual_dataset_correlacao.R
\name{eda_visual_dataset_correlacao}
\alias{eda_visual_dataset_correlacao}
\title{Gráfico de correlação}
\usage{
eda_visual_dataset_correlacao(dados, metodo = "pearson")
}
\description{
Gráfico de correlação
}
