% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_visual_dataset_densidade.R
\name{eda_visual_dataset_densidade}
\alias{eda_visual_dataset_densidade}
\title{Visualizar densidade de variáveis numéricas}
\usage{
eda_visual_dataset_densidade(dados, ncols = 2)
}
\description{
Visualizar densidade de variáveis numéricas
}
