% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_visual_dataset_disper.R
\name{eda_visual_dataset_disper}
\alias{eda_visual_dataset_disper}
\title{Gráficos de dispersão 2x2 usando ggplot2}
\usage{
eda_visual_dataset_disper(dados, ncols = 2)
}
\description{
Gráficos de dispersão 2x2 usando ggplot2
}
