% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_num.R
\name{fig_num}
\alias{fig_num}
\title{Gr치fico univariado para vari치vel quantitativa}
\usage{
fig_num(dados, coluna, fill = "steelblue", bins = 30, alpha = 0.8, ...)
}
\description{
Gr치fico univariado para vari치vel quantitativa
}
