% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_analise_descritiva.R
\name{relatorio_analise_descritiva}
\alias{relatorio_analise_descritiva}
\title{Resultados para relat칩rio de an치lise descritiva}
\usage{
relatorio_analise_descritiva(dados)
}
\arguments{
\item{dados}{dataset}
}
\description{
Resultados para relat칩rio de an치lise descritiva
}
