% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_relatorio.R
\name{sup_relatorio}
\alias{sup_relatorio}
\title{Saidas automaticas para listas (Fonte: Bel Carnielli)}
\usage{
sup_relatorio(lista)
}
\arguments{
\item{lista}{objeto da classe list()}
}
\description{
Saidas automaticas para listas (Fonte: Bel Carnielli)
}
