% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_rotular.R
\name{sup_rotular}
\alias{sup_rotular}
\title{Função para adicionar rotulos (labelled) a um dataset}
\usage{
sup_rotular(dados, rotulos)
}
\description{
Função para adicionar rotulos (labelled) a um dataset
}
