% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_dataset_descritiva_gts.R
\name{tabela_dataset_descritiva_gts}
\alias{tabela_dataset_descritiva_gts}
\title{Tabela descritiva usando gtsummary}
\usage{
tabela_dataset_descritiva_gts(dados)
}
\description{
Tabela descritiva usando gtsummary
}
