% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_resumo.R
\name{tbl_resumo}
\alias{tbl_resumo}
\title{Tabela de estatísticas descritivas em colunas usando gtsummary
https://site.statplace.com.br/blog/gtsummary-tabelas-de-resumo-prontas-para-publicacao/}
\usage{
tbl_resumo(dados, digits = NULL, missing = "ifany")
}
\description{
Tabela de estatísticas descritivas em colunas usando gtsummary
https://site.statplace.com.br/blog/gtsummary-tabelas-de-resumo-prontas-para-publicacao/
}
