% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes_tabelas.R
\name{testes_tabelas}
\alias{testes_tabelas}
\title{Gerar tabela de testes usando o gtsummary}
\usage{
testes_tabelas(dados, resposta, separado = TRUE)
}
\arguments{
\item{dados}{objeto de classe data.frame() ou tibble()}

\item{resposta}{variável}

\item{separado}{determina se será uma tabela unica ou tabekas separadas}
}
\description{
Gerar tabela de testes usando o gtsummary
}
