% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_sobre_dados.R
\name{texto_sobre_dados}
\alias{texto_sobre_dados}
\title{Texto descritivo de conjunto de dados para relatórios}
\usage{
texto_sobre_dados(dados, nome)
}
\description{
Texto descritivo de conjunto de dados para relatórios
}
