% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{visual}
\alias{visual}
\title{Visualização gráfica das variáveis}
\usage{
visual(dados)
}
\arguments{
\item{dados}{conjunto de dados}
}
\description{
Retorna plot com histogramas para variaveis numéricas e barras para categoricas
}
