% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_dataset_descritiva.R
\name{visual_dataset_descritiva}
\alias{visual_dataset_descritiva}
\title{Visualização de dados para análise descritiva}
\usage{
visual_dataset_descritiva(dados, cor = "cyan3")
}
\description{
Visualização de dados para análise descritiva
}
