% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_densidade.R
\name{visual_densidade}
\alias{visual_densidade}
\title{Visualizar densidade de variáveis numéricas}
\usage{
visual_densidade(dados, ncols = 2)
}
\description{
Visualizar densidade de variáveis numéricas
}
