% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_var_info.R
\name{dataset_var_info}
\alias{dataset_var_info}
\title{Informações relevantes das variáveis para testes estatísticos}
\usage{
dataset_var_info(dados, label = "na")
}
\description{
Informações relevantes das variáveis para testes estatísticos
}
