% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_visual_dataset_cat.R
\name{eda_visual_dataset_cat}
\alias{eda_visual_dataset_cat}
\title{Visualizar variáveis categoricas}
\usage{
eda_visual_dataset_cat(dados, resposta = NULL, ncols = 2, pos = "stack")
}
\description{
Visualizar variáveis categoricas
}
